Refactors frequency_distribution
`Array#to_h` is not available in Ruby <2.1, so I refactored with `Hash[]`. But just draping it around the contents of `frequency_distribution` made an already heavy method really hard to understand. Breaking it up into the two helper methods `order` and `raw_frequency` seems to make this easier to follow.