Fixes of_a_kind
The `of_a_kind` method was not counting of-a-kinds that were contained in greater of-a-kinds, e.g., `of_a_kind 2` would return 0 for rolls such [1, 1, 1, 2, 3] and [6, 6, 6, 6, 6].

Detecting `die_count >= kind_count` fixes this. I also added an `if_none` lambda to avoid returning `nil` so we can continue the calculation naturally.