Ensures two_pairs only scores distinct pairs
The rules for KataYahtzee are not explicit on whether rolls with 4 or 5 of the same die value are considered eligible for scoring. Since a kahtzee is explicitly not a full house, I've decided that `two_pairs` will only score distinct pairs — in all other cases, the score for this category will be 0.